More generalization for aggregation functions, allowing minimal logic requirements for new functions to be added
